pyproject.toml
[project]
name = "ai-powered-fitness-coach"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-openai>=0.3.30",
    "langchain-pinecone>=0.2.0",
    "pinecone-client>=6.0.0",
    "pydantic>=2.11.7",
    "pypdf>=6.0.0",
    "sqlalchemy>=2.0.43",
    "psycopg2-binary>=2.9.0",
    "uvicorn>=0.30.0",
    "python-multipart>=0.0.6",
    "streamlit>=1.48.1",
]
